+++ modules/Grades/Grades.php
@@ -550,7 +550,7 @@ foreach ( (array) $assignments_RET as $id => $assignment )
 
 		$optgroup = $types_RET[$assignment[1]['ASSIGNMENT_TYPE_ID']][1]['TITLE'];
 
-		$assignment_select .= '<optgroup label="' . htmlspecialchars( $optgroup ) . '">';
+		$assignment_select .= '<optgroup label="' . AttrEscape( $optgroup ) . '">';
 	}
 
 	$assignment_select .= '<option value="' . $id . '"' .